+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The most recent changes are on top, in each type of changes category.
 
 ### Added
 
+- E2E tests for Settings [#453])(https://github.com/cyfronet-fid/sat4envi/issues/453)
+
 ### Changed
 
 ### Fixed